refactor(core): move overlay movement logic to dedicated module
